requirements.txt
requests==2.31.0
pytesseract==0.3.10
Pillow==10.1.0
